WIP, discovered I need to find header within a chunk, not just p-1

+++ hw3/hw3.h
@@ -13,7 +13,7 @@ size_t* isPtr(size_t* p);
 int blockMarked(size_t* b);
 int blockAllocated(size_t* b);
 void markBlock(size_t* b);
-int length(size_t* b);
+long length(size_t* b);
 void unmarkBlock(size_t* b);
 size_t* nextBlock(size_t* b);
 void mark(size_t* p);
